+++ CHANGELOG.md
@@ -7,10 +7,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.0.0] - 2026-05-04
+
+### Added
+
+* V14 support
+* Rest button in a camping sheet can now be dragged onto the macro bar to rest quickly
+
+### Removed
+
+* Removed Teleporter Pair macro since this is now shipped in core
+* V13 Support
+
 ### Changed
 
+* Deleting a settlement from the kingdom sheet now offers an option to delete the scene as well
+* Settlement scenes now have a button to create blocks on the scene canvas
 * Infrastructure structure tokens do not increase settlement level anymore when placed on a settlement block
 * Mirror non kingmaker module settlement background images to align with the official module ones
+
+### Fixed
+
+* Fix DC reduction of lumber structures when setting is toggled
+
+
 
 ## [6.0.0-alpha1] - 2026-05-03